feat(type table): free the circular dependencies in `tag_core`, extract `PageElement`
