+++ src/bisque/typing/__init__.py
@@ -1,0 +1,3 @@
+from .tabulation import TypeTableBase
+
+__all__ = ["TypeTableBase"]
